refactor(auth): use descriptive file and method names
