+++ src/views/auth/RegisterView.vue
